CyberSecurity/Ataque/evil_jwt_force/requirements.txt
pyjwt
requests
beautifulsoup4
python-nmap
tk
